requirements.txt
python-sat==0.1.7.dev15
PyYAML==6.0
requests==2.22.0
termcolor==1.1.0
